Exercise 1: Complete
What I did with my exercise is that I wanted my circle to stop moving when the mouse is clicked and resume movement when the mouse is released.

However, I wanted my circle to show the emotion of fear by turning yellow and have the background go black when the mouse is pressed to show the lights being turned off.

The Cricle will turn yellow in fear and Qustion what's going on or evern you.

I achived the cricle assking a qusition thoguh the text function, and the String varible + an array that Victor kindly showed me how to do.

The Cricle will qustion when the screen is clicked and everthing will return to normal when the mouse is released.

PLEASE NOTE!!! all new lines of code I added will have a **at the bigging of the comments and code I have edited will have *at the start.

P.S sorry if this seem a little more then you asked, I was just having fun with  the exercise.