Exerises02: Done!
The changes I made was that:

I caused the background static to change size and color when the ball touches the paddle.

The ball flashes red when touching the paddle.

when touching the top of the screen I have the ball change speed.

I added a ballTop function that changes the ball's speed when the ball touches the top of the screen.