Exercise05: Done!
Rabbit Stair down!

what I did was make to bunny's have a consent stairdown with each other. they walk and move but are always facing each other.

I did this by using pushMatrix(); and popMatrix();

I first created the image and then translated it's scacle to -1 so that the image flips like a mirror.

 I also have an anmation for the bunny which works useing arrays and sin and maps.

When they are not moving they have an idle anmation

when they are moving they have an running anmation.

Aslo you contorl the bunnys using the below keys to move around and off the sreen.

W = Up
D = Down
A = left
D = Right