Exercise6:Complete!
I set the camera to look for the redest thing in the room then change the boncer's color deeping on the amount of red found.

I borrowed the code in the perstion side to help me out.

 for example how things work is that

if their is no red on the screnn then the boncers are black.

if somthing red is found but it's a little redest then give the boncers a little bit of red.

 If somthing is fully red that is found then set the boncers to the color red.

I also change the eslpse that tracks the redest obect into a rect and had it to chagne on the redness found as well. this was to make the tracker blind in with the boncer but not look to much like the boncer so that you can still see the redesst thing being found.