Exercise7:Done
What I did was borrow the mic vloume control from the slides. I mad the program yell out at you when the vloume us to high. so if you make to much nose the program is going to yell at you!

Also I used that I found online from a character from the Fire emablem Awaking game by Nintendo. The voice is tge chraracter Chrom but his voice does not sound the same when the program uses his voice clips.  There kind of high picth but I didn't do anyhting to fex this becuase this error is works in a way that I like it. I like it becuase the high picth voice suit the computers way of yealing at you.  It sounds mad but also cute which is way I don't want to chnge it.