Final Progect DONE!!
This is the last Commit before the deadline. Currently as I typing this it is the 12 of decmaber and tommrow is the due date... so this is one day head of it's self.  (im just saying this so you know I headed this in on time, not to be fancy).

Anyway there are 2 games: KeyRace and KeyButDon'tSpake

KeyRace is a basic lap around the keyboraed in loops which both players must complete 20 laps.

KeyButDontSpake: is a game where you must press down a set number keys before you or your buddie yells at the game to reset the game.

Have a good holiday Pippin!